Added dev/prod environment to grunt and express
When "dev" is passed as the only argument to the server, the app runs in
development mode and sets `dev = true`.

Now uses this dev arg to set up the static assets depending on
environemnt. In dev, adds ../app and ../bower_components -- for
production, just uses the built ../dist folder for static assets.

Also added a grunt command `grunt runprod` that builds the app for
production and runs the server in production mode. This is probably just
for testing the prod environment, since you don't need grunt after
building for real!